pyproject.toml
[tool.poetry]
name = "web-service-course-project"
version = "0.1.0"
description = ""
authors = ["Artem Deryavskiy <strelok.onaga@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "web_service_course_project"}]

[tool.poetry.dependencies]
python = "3.10.4"
fastapi = "^0.85.0"
pydantic = "^1.10.2"
pytest = "^7.1.3"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
black = "^22.8.0"
httpx = "^0.23.0"
pytest-httpx = "^0.21.0"
pytest-asyncio = "^0.19.0"
pytest-unordered = "^0.5.1"
pytest-mock = "^3.8.2"
pytest-cov = "^3.0.0"
codecov = "^2.1.12"
asgi-lifespan = "^1.0.1"
requests = "^2.28.1"
grpcio-tools = "^1.49.1"
Jinja2 = "^3.1.2"
Flask = "^2.2.2"
Werkzeug = "^2.2.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
